+++ cfg/cfg.go
@@ -23,12 +23,12 @@
 //     - node_modules
 //     import:
 //     - package: gopkg.in/yaml.v2
-//     - package: github.com/andrewlunde/vcs
+//     - package: github.com/Masterminds/vcs
 //       version: ^1.2.0
 //       repo:    git@github.com:Masterminds/vcs
 //       vcs:     git
 //     - package: github.com/urfave/cli
-//     - package: github.com/andrewlunde/semver
+//     - package: github.com/Masterminds/semver
 //       version: ^1.0.0
 //
 // These elements are: